ToyRepository 만들어 사용하라
ToyService에서 관리하던 toy 객체들의 책임을
ToyRepository로 분리합니다.
